修改worker加载方式 修改脚本类型为module
