+++ src/main/workers/coding.worker.ts
@@ -1,4 +1,5 @@
-import { encryptFF, decryptFF, encodeFF, decodeFF } from './utils/coding';
+import { compose } from 'redux';
+import { encode, decode } from 'js-base64';
 import registerPromiseWorker from 'promise-worker/register';
 
 type RecieveModule = 'encryptFF' | 'decryptFF' | 'encodeFF' | 'decodeFF';
@@ -20,3 +21,31 @@ registerPromiseWorker(async (params: WorkerRecieveParams) => {
       return decodeFF(params.data);
   }
 });
+
+function encodeFF(content: any) {
+  const ffprotocol = 'ff://'; // FF协议
+  return `${ffprotocol}${encode(JSON.stringify(content))}`;
+}
+
+function decodeFF(content: string) {
+  const ffprotocol = 'ff://'; // FF协议
+  try {
+    const protocolLength = ffprotocol.length;
+    const protocol = content.slice(0, protocolLength);
+    if (protocol !== ffprotocol) {
+      throw Error('协议错误');
+    }
+    const body = content.slice(protocolLength);
+    return JSON.parse(decode(body));
+  } catch (error) {
+    return null;
+  }
+}
+
+function encryptFF(content: string) {
+  return compose(encode, encodeFF)(content);
+}
+
+function decryptFF(content: string) {
+  return compose(decodeFF, decode)(content);
+}
