--- src/main/workers/utils/index.ts
@@ -1,1 +1,0 @@
-export * as Coding from './coding';
